requirements.txt
waveshare-epaper>=1.1.1